Move capability implementation to separate namespace
