+++ src/Cake.Issues.PullRequests.Tfs.Tests/Capabilities/CommentExtensionsTests.cs
@@ -1,5 +1,6 @@
-﻿namespace Cake.Issues.PullRequests.Tfs.Tests
+﻿namespace Cake.Issues.PullRequests.Tfs.Tests.Capabilities
 {
+    using Cake.Issues.PullRequests.Tfs.Capabilities;
     using Cake.Issues.Testing;
     using Microsoft.TeamFoundation.SourceControl.WebApi;
     using Shouldly;
