+++ src/Cake.Issues.PullRequests.Tfs/Capabilities/CommentThreadStatusExtensions.cs
@@ -1,4 +1,4 @@
-﻿namespace Cake.Issues.PullRequests.Tfs
+﻿namespace Cake.Issues.PullRequests.Tfs.Capabilities
 {
     using Microsoft.TeamFoundation.SourceControl.WebApi;
